feat(drawings): Moved core drawing primitives to separate directory
