+++ src/lib/ui/app/Chart/drawing-tools/_core/pane-view.ts
