+++ src/lib/ui/app/Chart/drawing-tools/_core/renderer.ts
