+++ src/lib/ui/app/Chart/drawing-tools/fib-retracement/primitive.ts
@@ -1,6 +1,6 @@
 import type { TPoint } from '../types.js'
 
-import { DrawingPrimitive } from '../primitive.js'
+import { DrawingPrimitive } from '../_core/primitive.js'
 import { FibRetracementPaneView } from './pane-view.js'
 
 export default class FibRetracementPrimitive extends DrawingPrimitive<'fib_retracement'> {
