+++ src/lib/ui/app/Chart/drawing-tools/rectangle/renderer.ts
@@ -2,6 +2,8 @@ import type { CanvasRenderingTarget2D } from 'fancy-canvas'
 import type { IPrimitivePaneRenderer } from '@santiment-network/chart-next'
 import type { TViewPoint } from '../types.js'
 
+import { positionsBox } from '../_core/renderer.js'
+
 export class RectanglePaneRenderer implements IPrimitivePaneRenderer {
   _p1: TViewPoint
   _p2: TViewPoint
@@ -29,31 +31,4 @@ export class RectanglePaneRenderer implements IPrimitivePaneRenderer {
       )
     })
   }
-}
-
-/**
- * Determines the bitmap position and length for a dimension of a shape to be drawn.
- * @param position1Media - media coordinate for the first point
- * @param position2Media - media coordinate for the second point
- * @param pixelRatio - pixel ratio for the corresponding axis (vertical or horizontal)
- * @returns Position of of the start point and length dimension.
- */
-export function positionsBox(
-  position1Media: number,
-  position2Media: number,
-  pixelRatio: number,
-): BitmapPositionLength {
-  const scaledPosition1 = Math.round(pixelRatio * position1Media)
-  const scaledPosition2 = Math.round(pixelRatio * position2Media)
-  return {
-    position: Math.min(scaledPosition1, scaledPosition2),
-    length: Math.abs(scaledPosition2 - scaledPosition1) + 1,
-  }
-}
-
-export interface BitmapPositionLength {
-  /** coordinate for use with a bitmap rendering scope */
-  position: number
-  /** length for use with a bitmap rendering scope */
-  length: number
 }
